requirements.txt
karton-core>=5.0.0
requests
sanic
schema
